Split SSL Quic test
